ci(test): Merge test and checks (#3233)
# Motivation

There is some concurrency error with the status check `may-merge` of our
CIs. Apparently, if one of the status check is positive, and the rest is
not, the PR is cleared to merge.

We had a few instances of it in the past, but here is a practical
example with PR #3207 : in the last commit before merge, [the workflow
for the checks
failed](https://github.com/dfinity/oisy-wallet/actions/runs/11576857312),
but the one for the tests passed. So the latest may-merge superseded the
first one, and the PR was allowed to merge.

![Screenshot 2024-10-30 at 09 28
58](https://github.com/user-attachments/assets/d5830ae7-1915-4f59-a55a-bcc01e88eea8)

![Screenshot 2024-10-30 at 09 29
27](https://github.com/user-attachments/assets/37c963cf-b2f2-4b57-b797-224db14a76d5)




To try and solve it, we first merge the two workflows, so that they
share the same status check `may-merge`