fix(frontend): Quickfix sorting for Token Groups (#3262)
# Motivation

It may happen that after the grouping, a particular group results with a
balance that is greater than another group/token above it. So the
sorting could be wrong:


![image](https://github.com/user-attachments/assets/b9398bf4-9ecf-4eba-a4e1-e3cb9ffd9658)

We are going to refactor the grouping util, but not in this PR.

For now, it was requested a "quickfix" to at least have the correct
sorting. To achieve that, we sort the final list of `[tokens + groups]`
by balances (token and USD).

It is guaranteed that the only sorting criteria that is being mixed is
the one concerning the balances, because a group could only have a
greater (or same) balance than the single tokens it was born from.

The other sorting criteria (name, network.name) can be ignored because
the resulting position given by them will be maintained during grouping.