fix(e2e): Make ButtonHero loaded when exchange disabled (#3350)
# Note

This is a duplication of #3314 , because I changed my account.

# Motivation

The buttons in the Hero were permanently loading when the Exchange is
are not enabled (never initialized). That mostly happens in test
environments, such as the one used for E2E test. We can see the
artifacts of [workflow E2E test ran on
main](https://github.com/dfinity/oisy-wallet/actions/runs/11662656812):

![Screenshot 2024-11-04 at 18 43
08](https://github.com/user-attachments/assets/7017273b-8663-47bd-823e-db03b3a5f794)

To solve it, we disable the exchanges.

# Changes

- Create env file for E2E tests with disabled exchanges.
- Add the file to playwright configs.

# Tests

I run the [screenshots workflow on this
branch](https://github.com/dfinity/oisy-wallet/actions/runs/11684626400)
and it worked.