refactor(frontend): Simplify props of Hero (#3347)
# Motivation

The props given to `Hero` are somehow a bit confused since derived from
routing. They were ok before, since the `Hero` was used in all routes.
With the latest changes, the scope of the `Hero` was reduced to only the
main page (tokens) and the transaction page.

To clean-up a bit we refactor the Hero to be a bit more obvious. In
another PR, we will probably merge `Hero` and `HeroContent`.

# Changes

- Use navigations utils directly in `HeroContent` and `+layout`
- Remove deprecated props from Hero.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>