refactor(frontend): use common component for TransactionStatusComponent (#3397)
# Motivation

We want to try and create more common components for the Transaction
components. One that is already ready to use is the
TransactionStatusComponent. So we replace `TransactionPending` with
`TransactionStatusComponent`.

The only difference between them is that the new one consider even the
`unconfirmed` status (proper of BTC transactions) beyond the `pending`
status.