feat(frontend): split ic-wallet.scheduler (#3464)
# Motivation

We will need another scheduler to support tokens without index canister.
Therefore, we split the existing `ic-wallet.scheduler` in two. An
abstract class for the start, stop etc. features and the existing
implementation to `ic-wallet-transactions.scheduler? .

# Notes

No features or code was implemented here. It's a feature but can also be
considered as refactoring.

# Changes

- Split code as described
- Use new "child" class in workers