feat(backend): Add topping up API (#3469)
# Motivation
We need to automatically top up the oisy cycles ledger account.

We do this by having an API method on the backend canister that tops up
the account. The method can be called by canister controllers or
automatically on a timer.

# Changes
- Define (but do not implement) an API method to top up the ledger
canister.
  - Implementation is omitted to keep the PRs small.
- Update the candid file and bindings.

# Tests
CI has an existing test to verify that the bindings are up to date.