feat(frontend): enable usdt usdc ckusdt by default but let user disable them (#3337)
# Motivation

USDT, USDC and ckUSDT should by default be enabled, but not locked, so
users can disable them. This MR extends the
`mapDefaultTokenToToggleable` which currently handles the toggling of
ICRC default canisters. It introcuces constants for ICRC and ERC20 to
track the suggested tokens. For suggested tokens, the user preference of
enabling/disabling has a higher priority.

# Changes

- introduce constants for suggested canisters / tokens
- enhance mapping default tokens to toggleable

# Note

A further MR will address the sepolia tokens to be handled in the same
manner.

# Tests

Fresh user:


https://github.com/user-attachments/assets/9239a86b-09eb-4407-aaf6-24df618314af


Existing user:


https://github.com/user-attachments/assets/d291dd6a-fd20-40cf-bbf3-9ad5f5949f49

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>