feat(frontend): util to format a timestamp to a normalized date (#3526)
# Motivation

For grouping the transactions by date, we create an util that formats a
timestamp to a normalized date, for example, using today as reference:

- `Today` if it's today, i.e. `13 November 2024`
- `Yesterday` if it's yesterday, i.e.`12 November 2024`
- `17 June` if it's in the same year
- `31 December 2023` if is in the past years

However , we provide an optional param that is the date of reference,
since we want to maintain the same reference if used in a loop. Maybe
over-engineered, but it feels more correct.

# Note

The final scope is to present the transactions in smaller groups, by
date, for example (this is just a mockup):

![Screenshot 2024-11-13 at 17 09
24](https://github.com/user-attachments/assets/f5de23f8-67ba-421d-befe-422354e84947)
![Screenshot 2024-11-13 at 17 10
09](https://github.com/user-attachments/assets/2cd780cb-f85f-4a4d-aca4-6f83407167b2)


# Tests

Included some tests.