feat(frontend): util to map unified transactions list (#3458)
# Motivation

Following up PR #3449 , we implement the util to create the list of
unified transactions, starting with BTC transactions.

# Changes

- Move the util from inside the `AllTransactionsList` component to a
specific module.
- Use the existing transaction store for BTC and map their component.

# Tests

Created some additional tests for both the util and the component. Here
are the result loading a BTC address that has a transaction. Please note
that the UI will be modified later.


![Screenshot 2024-11-12 at 15 24
56](https://github.com/user-attachments/assets/bc7cb738-cb2e-4def-93e7-b82f7fd7c663)
![Screenshot 2024-11-12 at 15 25
03](https://github.com/user-attachments/assets/b396054b-5b90-4201-8a1f-523acd716169)