feat(devops): Manually deploy breaking changes via CI (#3546)
# Motivation
If an API changes, dfx will typically ask the user whether they really
want to deploy. In CI, we need a way of saying "yes", but probably don't
want this to be the default.

There is an existing checkbox for forcing the backend deployment to take
place, no matter what. It seems reasonable to use `--yes` when that is
selected, so we do not need a new input.

# Changes
- If `force-backend` is selected in the CI deployment task, deploy with
`--yes` and `--upgrade-unchanged`.

# Tests
Executed the action by:
- Disabling the check that allows deploying only from main.
- Running the action with: `gh workflow run deploy-to-environment.yml
--ref "yes" --field network=staging --field canister=backend --field
force-backend=true`
- Re-enabling the main check.