feat(frontend): add Ethereum transactions to unified list (#3547)
# Motivation

Following PR #3458 , we include Ethereum transactions in the unified
list. It simply uses the same logic of the BTC transactions.

# Note

A few notes:
- For now, we include Sepolia, because it is easy for testing; I
included a TODO to filter after the feature is complete.
- We do not test `mapEthTransactionUi` (it is tested elsewhere), so we
provide undefined props for it. This function is used mostly for
defining the ID and the transaction type, that is not really relevant
for the tests of this PR.
- We still don't treat sorting and grouping in this PR, we are just
focusing on unifing.

# Changes

- Create util to verify Sepolia networks.
- Add ETH transactions store in mapping.

# Tests

I included more tests, both for the util and the component.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>