refactor(frontend): remove token derived stores from Amount component (#3580)
# Motivation

The component `Amount` was using a derived of `token` store. However, it
is now used in a context where the token may change and it is not
defined, for example, it is mounted on each transaction on the unified
list.

So, I thought of giving the adequate properties instead of using the
store. But that means that we need to propagate such properties.

I thought of creating a context too, but it will change for each Amount
component mounted in the unified transaction list, so not a possible
solution for now.

# Changes

- Add prop `decimals` to component `Amount`.
- Remove derived store dependency from `Amount`.
- Adapt component `Balance` that is using `Amount`.
- Adapt component `Transaction` that is using `Amount`.
- Adapt all transaction components for each network, providing at the
highest level the token as token with fallback (that was what it was
used before inside the `Amount` component, so no real change).
- We force the token to be non-optional for convenience, since in the
future, we will be removing the `tokenWithFallback` store and we will
need these components to have a defined token.

# Tests

No changes in local replica.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>