refactor(frontend): move hide info utils to common folder (#3607)
# Motivation

Since we are going to re-use them, we move the hide info utils to the
common `lib` folder. And we add tests.

# Note

I am not using zod for the keys list because I am replicating the way we
do for modals, defining explicit string types. I am not 100% sure that
zod enforces the strings in the same way.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>