feat(frontend): add burn icrc token from json (#3523)
# Motivation

We add the additional icrc token "BURN".

# Changes

- adjust type reference
- add token to json
- merge data into icrc tokens

# Tests

In discussion with @AntonioVentilii-DFINITY I aimed for manual testing
because testing this functionality in a unit test would include multiple
stores and a deeper knowledge of how they work together.

Acceptance criteria:
User w/o the token: user does not see the token in the list, and sees
the token in the list settings and can enable it

2650331


https://github.com/user-attachments/assets/12139ae4-a4ee-4d60-94d7-6c152a63856d


User w/ the token, a balance, active: User still sees the token and the
balance in the list. and only once in “manage list” and can enable it

2650332


https://github.com/user-attachments/assets/7baad085-ecca-4f3d-a841-726c512cc81e



![image](https://github.com/user-attachments/assets/eed64e88-ade9-4fd1-bcc9-e0f06db4f44a)


User w/ the token, a balance, inactive: User does not see the token in
the list. and only once in “manage list” and can enable it

2650334



https://github.com/user-attachments/assets/e4abf017-b0f1-456e-b31f-792071a58883

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>