refactor(frontend): simplify restRequest (#3664)
# Motivation

In PR #3637 , we created an util to send REST requests with a certain
amount of retries, before failing.

However it was over-engineered, and some logic didn't quite make sense.
For example, it doesn't make sense to have `onSuccess` and `onError`
return values. This should be handled on the caller side.

So, we greatly simplify the function and we make it even nested (more
functional).

# Changes

- Remove `onSuccess`: the function will return the request results, if
successful.
- Remove `onError`: the function will throw directly the error if it
exceeded the max attempts.
- Create a sub-fuction that is called nested inside the main function.

# Tests

Adapted the tests, just removing the parts concerning `onSuccess` and
`onError`.