feat(frontend): update utxos fee store and context (#3728)
# Motivation

After testing the conversion flow, I noticed that UTXOS fees are being
re-fetched when switching between WizardModal steps. After debugging the
whole flow, I noticed that BtcConvertTokenWizard is re-mounted every
time a step changes (it actually happens in all places where we use
WizardModal). To fix it for the BTC -> ckBTC (where fetching utxos takes
ages), I did the following things:

1. Refactored a bit utxos fee store and context, so the related is store
is initialised outside of Svelte. It actually simplified all unit tests
where it's used.
2. Started to save the inputed amount for which utxos were already
fetched and to compare it with the new value to prevent re-fetch if it's
the same (additional unit tests are provided).

Declaring utxos-fee context outside of WizardModal would also solve the
problem, but the parent would need to know too much data that is set in
children components (sendAmount, amountErrors, etc).