ci(e2e): added config to handle carousel (#3688)
# Motivation

Since the implementation of the Dapps Carousel, the e2e tests were
inconsistent. Due to the carousel automatically switching slides there
was a possibility for screenshot 1 to be on slide x and screenshot 2 to
be on slide y. For a consistent and useful test this needs to be
changed.

# Changes

- added config that whenever the test for the homepage is run, it will
check the carousel slides and always take the screenshot on the last
slide

# Tests

tested multiple times on local instance.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Antonio Ventilii <antonio.ventilii@dfinity.org>
Co-authored-by: AlessandroBonomo <bonomo.alessandro@yahoo.de>