feat(frontend): load utxos fee on convert modal open (#3848)
# Motivation

In this PR, 2 updates on UtxosFeeContext have been implemented:

1. We now start loading utxos fee directly when user opens the convert
modal. To do that, we pass to the endpoint a default amount specified in
the code. The fees are usually the same (no matter what "amount" is
provided) for a long period of time, therefore it won't lead to an
incorrect value being displayed.
2. After we got a fee value, we don't instantly reset the store if user
provides a new amount. Instead, we continue displaying the old value and
fetch the new fee in the background.