feat(frontend): extract User Profile loader (#3844)
# Motivation

The user profile will be used not only for the Settings page, but for
other components through OISY (for example the carousel settings, see PR
#3770 ).

So, we first extract it into a loader component at the top level layout.

However, we still need to force a refresh sometimes, so we make it
listen to a specific event.

# Changes

- Extract component `LoaderUserProfile`.
- Load user profile on `onMount`.
- Create custom event `oisyRefreshUserProfile`.
- Use the event to re-load the user profile in the new component.
- Adapt `Settings`.

# Tests

Included some unittest for the new component.