test(frontend): Tests for certified store init (#3873)
# Motivation

We create some base tests for `initCertifiedStore` and
`initCertifiedSetterStore`

Note that we are not creating tests for the `update` function since it
is replicated as it is from `writable` constructor of `svelte`. Nor we
create a test for `reset` in `initCertifiedSetterStore` since it is a
copy of the same function inherited from `initCertifiedStore`.