feat(frontend): function to remove slide for Carousel (#3869)
# Motivation

The component Carousel is not reactive to its children in `<slot />` (it
will be possible only in Svelte 5). But we need to have a way to
"refresh" the slides. Our first use-case is to remove a specific slide.

Since the component modify its children on mounting, we create a
function that can be called outside the component to remove a slide at a
particular index position.

# Changes

- Add function `removeSlide` to `Carousel` and export it.
- Re-initialise the `Carousel` once the slide is removed.
- Stop timer if there is only one slide.
- Add transitions.

# Tests

We still don't have unittest with custom components (like GIX
Components), nor an E2E test for this. However, I simulated a random
removal of one of the slides from the parent of Carousel.

```typescript
// every 5 seconds remove a dapp from the carousel
const interval = setInterval(() => {
  if (nonNullish(carousel)) {
    const dapp = dappsCarouselSlides[Math.floor(Math.random() * dappsCarouselSlides.length)];
    
    console.log('removing', dapp?.id);
    const idx = dappsCarouselSlides.findIndex((d) => d.id === dapp?.id);
    
    carousel.removeSlide(idx);
  }
}, 5000);
```


The results are:



https://github.com/user-attachments/assets/2acb4440-ba94-4879-ace0-c368035aa4c1

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>