fix(frontend): load Carousel when missing settings in user profile (#3893)
# Motivation

The carousel was not being loaded for old accounts, because
`userSettings` was always `undefined`.

However, we should distinguish between not having loaded the user
profile (from which `userSettings` derives), and having a missing field
`settings` in the user profile (it happens for old accounts, since we
introduced the field recently).

# Changes

- Improve `userSettings` derived store to set `null` when the user
profile exists but it is missing the `settings` field.
- Adapt tests.
- Adapt `DapsCarousel`.

# Tests

Provided.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>