feat(frontend): derived store userProfileLoaded (#3895)
# Motivation

According to @peterpeterparker 's suggestion in
https://github.com/dfinity/oisy-wallet/pull/3893#pullrequestreview-2486419830
, we find a better way to distinguish if the user settings are not
available because the user profile is not loaded, or if they are missing
(old accounts for example).

The base difference is simply if the user profile is not `nullish`: if
it is not `nullish` and the settings are nullish, we are sure that the
`settings` are missing.

# Changes

- Create derived store `userProfileLoaded`.
- Use new store in `DappsCarousel`.

# Tests

Test created and adjusted.