style(frontend): fix modal behavior (#3859)
# Motivation

When a user opens a modal that does some frontend calculations like the
`Manage Tokens` modal and then clicks the browser's back button, the
modal disappears and the previous page is loaded. This behaviour is not
what we want. If the browser's back button is clicked while a modal,
such as `Manage Tokens`, is open, the user should see a default browser
confirmation message asking whether he wants to leave the current page
or stay on it.

# Changes

- user needs to confirm to navigate to another page if heavy
calculations are made in a modal

# Tests

**before:**


https://github.com/user-attachments/assets/fdb0539e-ee58-41ed-a5f5-9ce187435847



**after:**


https://github.com/user-attachments/assets/78f5d317-3c0b-4f1c-929e-6fa6113ce22c

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>