ci(frontend): merge auto-format in checks (#3894)
# Motivation

It would make sense to merge the auto-format CI into the frontend
checks, since they are already checking the format. So that we spare one
Ci and already fix the formatting if needed, during the checks.

Plus, it is quicker.

# Tests

I injected bad formatting in this same PR and it was fixed.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>