fix(frontend): Carousel glitch for last slide (#3920)
# Motivation

When an user closes a slide in the dApp, the backend is updated and the
user profile re-loaded. However it may happen that the two operations
are concurrent. So we could have a non-updated user profile.

That causes basically to not really hide the slide that the user wants
to close.

We avoid this glitch saving the closed dApps in temporary array that we
use to filter, together with the user settings.



### Before


https://github.com/user-attachments/assets/0eeac073-e12c-4bf6-a392-85fc0818cc7f

### After



https://github.com/user-attachments/assets/807ecb73-6f00-4f39-98b3-84a1a00bb100

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>