feat(frontend): listener templating (#3937)
# Motivation

Migration to Svelte v5 (#3929) involves replacing the ComponentType.
While updating the Listener, a minor TypeScript issue was identified. We
use a `svelte:component` that maps a token property, which may or may
not be present, and can sometimes be set or unset. This approach works
because the component is assigned to a state variable, but it is not
fully TypeScript-compliant. To address this, we can leverage the
template to correctly mount the listener.